+++ package/src/components/thumb/x.tsx
@@ -1,0 +1,65 @@
+import type * as DOM from "atomico/types/dom";
+
+import type { ComponentTypes } from "#/@types/component";
+
+import * as Atom from "atomico";
+
+import { useScrollCore } from "#/contexts/scrollcore";
+import { useComponentPropsSetter } from "#/hooks/props";
+import { useThumbHandlerX } from "#/hooks/thumb/x";
+
+const _ThumbX = () => {
+    const elRef: Required<Atom.Ref<DOM.AtomicoThis>> = Atom.useHost();
+
+    const {
+        x: { setHvThumb, scrollbarLength, scrollbarOffset },
+    } = useScrollCore();
+
+    Atom.useEffect((): void => {
+        setHvThumb(true);
+    });
+
+    const { onPointerDown } = useThumbHandlerX();
+
+    Atom.useEffect((): (() => void) => {
+        elRef.current.addEventListener("pointerdown", onPointerDown);
+
+        return (): void => {
+            elRef.current.removeEventListener("pointerdown", onPointerDown);
+        };
+    }, [
+        onPointerDown,
+    ]);
+
+    useComponentPropsSetter("thumbX");
+
+    Atom.useEffect((): void => {
+        const length: number = Number.isNaN(scrollbarLength)
+            ? 0
+            : scrollbarLength;
+
+        const offset: number = Number.isNaN(scrollbarOffset)
+            ? 0
+            : scrollbarOffset;
+
+        elRef.current.style.width = `${length}px`;
+        elRef.current.style.left = `${offset}px`;
+    }, [
+        scrollbarLength,
+        scrollbarOffset,
+    ]);
+
+    return (
+        <host shadowDom>
+            <slot />
+        </host>
+    );
+};
+
+/** Scrollbar thumb X type. */
+type ThumbXElement = ComponentTypes<typeof _ThumbX>;
+
+const ThumbX = Atom.c(_ThumbX);
+
+export type { ThumbXElement };
+export { ThumbX };